moved retail module
