+++ core/domain/src/main/java/com/wansenai/dto/receipt/retail/QueryShipmentsDTO.java
@@ -10,7 +10,7 @@
  * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package com.wansenai.dto.receipt;
+package com.wansenai.dto.receipt.retail;
 
 import lombok.Data;
