+++ core/domain/src/main/java/com/wansenai/dto/receipt/retail/RetailShipmentsDTO.java
@@ -10,47 +10,41 @@
  * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package com.wansenai.vo.receipt;
+package com.wansenai.dto.receipt.retail;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import com.wansenai.bo.BigDecimalSerializerBO;
 import com.wansenai.bo.FileDataBO;
 import com.wansenai.bo.ShipmentsDataBO;
 import lombok.Builder;
 import lombok.Data;
 
 import java.math.BigDecimal;
-import java.time.LocalDateTime;
 import java.util.List;
 
 @Data
 @Builder
-public class RetailShipmentsDetailVO {
+public class RetailShipmentsDTO {
 
-    @JsonFormat(shape = JsonFormat.Shape.STRING)
+    private Long id;
+
     private Long memberId;
 
-    @JsonFormat(shape = JsonFormat.Shape.STRING)
     private Long accountId;
 
-    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "GMT+8")
-    private LocalDateTime receiptDate;
+    private String receiptDate;
 
     private String receiptNumber;
 
     private String paymentType;
 
-    @JsonSerialize(using = BigDecimalSerializerBO.class)
     private BigDecimal collectAmount;
 
-    @JsonSerialize(using = BigDecimalSerializerBO.class)
     private BigDecimal receiptAmount;
 
-    @JsonSerialize(using = BigDecimalSerializerBO.class)
     private BigDecimal backAmount;
 
     private String remark;
+
+    private Integer status;
 
     private List<ShipmentsDataBO> tableData;
