+++ core/domain/src/main/java/com/wansenai/vo/receipt/retail/RetailShipmentsVO.java
@@ -1,16 +1,4 @@
-/*
- * Copyright 2023-2033 WanSen AI Team, Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance
- * with the License. A copy of the License is located at
- *
- * http://opensource.wansenai.com/apache2.0/
- *
- * or in the "license" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES
- * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
- * and limitations under the License.
- */
-package com.wansenai.vo.receipt;
+package com.wansenai.vo.receipt.retail;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
@@ -27,7 +15,7 @@ import java.time.LocalDateTime;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class RetailRefundVO {
+public class RetailShipmentsVO {
 
     @JsonFormat(shape = JsonFormat.Shape.STRING)
     private Long id;
@@ -49,7 +37,7 @@ public class RetailRefundVO {
     private BigDecimal totalPrice;
 
     @JsonSerialize(using = BigDecimalSerializerBO.class)
-    private BigDecimal paymentAmount;
+    private BigDecimal collectionAmount;
 
     @JsonSerialize(using = BigDecimalSerializerBO.class)
     private BigDecimal backAmount;
